SqLVim v2.3.0, The Terminal update
The terminal update added a new terminal menu where you can multiple terminals running in the background. With the user of <leader>tf and the number you entered, you created a new floating terminal on the index you gave to the terminal. Keep in mind it works with numbers. NOT WITH STRINGS.

Check out the toggleterm.lua file for the keybinds and change how you see fit.