SqLVim v2.4.0, better dir update.
In this update I basically grabbed all the files and put them in either a shared dir or single dir. Now it will be easier to find the files you are looking for. The structure is not perfect now, and gitsigns.nvim is also missing. Don't worry this wil get fixed later today. Happy vimming!