SqLVim 2.7.0 Back to the roots update
To keep it short. I made my promise to clean up SqLVim but in a pretty obvious way. 

Almost all the plugins you use on a daily basis like bufferline, dressing and more are just fancy layers of abstraction. Yes they provide some form of customization but the downside is that a lot of these things can be done in vanilla vim already. Because of that I made the choice of removing some functionality like bufferline and dressing.

Fear not because the functionality that those plugins had can be mimicked using native vim buffers and tabs. 

But why would this be removed?

It was mostly related to performance. As stated the philosophy of SqLVim is being a lighweight neovim IDE. Yes bufferline made it feel more like a true IDE but came with the cost of it being extra resources to do what vim can already do in a fancier way. Same with dressing, yes it looks nice but the added features you get with dressing are just not enough for me to use it. 

In a perfect world we would only use native vim solutions for the fastest result, however this is just not possible. In the feature more plugins might get modified or changed to be better for performance and more open to vanilla vim use cases.